support for specification of schema in pandas extension `pg_copy_from`
in line with pandas, schema support is implemented as a separate
keyword, `schema=None`, supported only when an extant table is specified
as the `copy_from` source.

if the source `sql` is not recognized as an extant table, and `schema`
is specified, `TypeError` is raised, as previously with argument `columns`.

additional context is added to this error message, in the case that
`sql` **appears** to be a table reference; (however, we do not otherwise
attempt to logically distinguish between a table reference and a sql
expression based on the contents of the string).

resolves #23